+++ src/shared/components/palette/base-field/field-write.component.ts
@@ -2,10 +2,8 @@ import { Component, ComponentFactoryResolver, Injector, Input, OnInit, ViewChild
 import { AbstractControl } from '@angular/forms';
 import { plainToClassFromExist } from 'class-transformer';
 import { CaseEditPageComponent } from '../../case-editor/case-edit-page/case-edit-page.component';
-
 import { CaseField } from '../../../domain/definition';
 import { FormValidatorsService } from '../../../services/form';
-import { CaseEditPageComponent } from '../../case-editor/case-edit-page/case-edit-page.component';
 import { PaletteService } from '../palette.service';
 import { AbstractFieldWriteComponent } from './abstract-field-write.component';